backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.4

# LangChain (compatible versions)
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0

# Google Gemini
google-generativeai>=0.8.0

# Cohere Embeddings
cohere>=5.0.0

# Voyage AI Embeddings
voyageai>=0.2.3

# NumPy (vectorised similarity search)
numpy>=1.24.0

# Local Embeddings (BGE-M3) — optional, uncomment if needed
# sentence-transformers>=2.2.2
# torch>=2.1.0

# Vector DB - Qdrant
qdrant-client

# Telegram Bot
python-telegram-bot>=21.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.7.4
pydantic-settings>=2.10.1
aiofiles==23.2.1
orjson>=3.9.0

# HTTP Client
httpx>=0.27.0
pyTelegramBotAPI